Introduce abstract parent class for admins user commands
