+++ src/Sylius/Bundle/AdminBundle/Command/CreateAdminUserCommand.php
@@ -19,8 +19,6 @@ use Symfony\Component\Console\Attribute\AsCommand;
 use Symfony\Component\Console\Command\Command;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Output\OutputInterface;
-use Symfony\Component\Console\Question\Question;
-use Symfony\Component\Console\Style\SymfonyStyle;
 use Symfony\Component\Intl\Locales;
 use Symfony\Component\Messenger\Exception\HandlerFailedException;
 use Symfony\Component\Messenger\HandleTrait;
@@ -30,22 +28,15 @@ use Symfony\Component\Messenger\MessageBusInterface;
     name: 'sylius:admin-user:create',
     description: 'Create a new admin user',
 )]
-final class CreateAdminUserCommand extends Command
+final class CreateAdminUserCommand extends AbstractAdminUserCommand
 {
     use HandleTrait;
-
-    private SymfonyStyle $io;
 
     public function __construct(MessageBusInterface $messageBus, private string $defaultLocaleCode)
     {
         $this->messageBus = $messageBus;
 
         parent::__construct();
-    }
-
-    protected function initialize(InputInterface $input, OutputInterface $output): void
-    {
-        $this->io = new SymfonyStyle($input, $output);
     }
 
     protected function execute(InputInterface $input, OutputInterface $output): int
@@ -105,40 +96,6 @@ final class CreateAdminUserCommand extends Command
         $adminUserData['enabled'] = $this->io->confirm('Do you want to enable this admin user?', true);
 
         return $adminUserData;
-    }
-
-    private function createEmailQuestion(): Question
-    {
-        $question = new Question('Email');
-        $question->setValidator(function (?string $email) {
-            if (!filter_var($email, \FILTER_VALIDATE_EMAIL) || $email === null) {
-                throw new \InvalidArgumentException('The email address provided is invalid. Please try again.');
-            }
-
-            return $email;
-        });
-        $question->setMaxAttempts(3);
-
-        return $question;
-    }
-
-    private function createQuestionWithNonBlankValidator(string $askedQuestion, bool $hidden = false): Question
-    {
-        $question = new Question($askedQuestion);
-        $question->setValidator(function (?string $value) {
-            if ($value === null) {
-                throw new \InvalidArgumentException('The value cannot be empty.');
-            }
-
-            return $value;
-        });
-        $question->setMaxAttempts(3);
-
-        if ($hidden) {
-            $question->setHidden(true);
-        }
-
-        return $question;
     }
 
     private function showSummary(array $adminUserData): void
